feat: C# | add solution for problem #3

+++ CSharp/SolutionInfo.cs
@@ -3,7 +3,7 @@ namespace ProjectEuler
     public class SolutionInfo
     {
         public int ProblemNumber;
-        public int ProblemSolution;
+        public long ProblemSolution;
         public long ExecutionTimeInMs;
 
         public SolutionInfo(int problemNumber)
